+++ backend/CHANGELOG.rst
@@ -7,6 +7,7 @@ This file keeps track of all notable changes to license-manager-backend
 Unreleased
 ----------
 * Update docker-compose to use postgresql instead of postgres
+* Added search and sort to list endpoints.
 
 2.2.5 -- 2022-04-12
 -------------------